edit data type subjectId

+++ models/Subject.js
@@ -3,7 +3,7 @@ module.exports = (sequelize, dataTypes) => {
     "Subject",
     {
       subjectId: {
-        type: dataTypes.INTEGER(8),
+        type: dataTypes.INTEGER,
         primaryKey: true,
         allowNull: false,
       },
